+++ plenum/test/bls/helper.py
@@ -1,3 +1,5 @@
+from contextlib import contextmanager
+
 import base58
 import os
 
@@ -22,6 +24,19 @@ from plenum.test.pool_transactions.helper import sdk_send_update_node, \
 from stp_core.common.log import getlogger
 
 logger = getlogger()
+
+
+@contextmanager
+def update_validate_bls_signature_without_key_proof(txnPoolNodeSet, value):
+    default_param = {}
+    for n in txnPoolNodeSet:
+        config = n.bls_bft.bls_key_register._pool_manager.config
+        default_param[n.name] = config.VALIDATE_BLS_SIGNATURE_WITHOUT_KEY_PROOF
+        config.VALIDATE_BLS_SIGNATURE_WITHOUT_KEY_PROOF = value
+    yield value
+    for n in txnPoolNodeSet:
+        n.bls_bft.bls_key_register._pool_manager. \
+            config.VALIDATE_BLS_SIGNATURE_WITHOUT_KEY_PROOF = default_param[n.name]
 
 
 def generate_state_root():
@@ -275,6 +290,7 @@ def validate_proof_for_read(result, req):
                                             serialized=True)
     return valid
 
+
 def validate_proof_for_write(result):
     """
     Validates state proof
@@ -341,6 +357,26 @@ def validate_multi_signature(state_proof, txnPoolNodeSet):
     return _multi_sig_verifier.verify_multi_sig(signature,
                                                 value,
                                                 public_keys)
+
+
+def update_bls_keys_no_proof(node_index, sdk_wallet_stewards, sdk_pool_handle, looper, txnPoolNodeSet):
+    node = txnPoolNodeSet[node_index]
+    sdk_wallet_steward = sdk_wallet_stewards[node_index]
+    new_blspk, key_proof = init_bls_keys(node.keys_dir, node.name)
+    node_dest = hexToFriendly(node.nodestack.verhex)
+    sdk_send_update_node(looper, sdk_wallet_steward,
+                         sdk_pool_handle,
+                         node_dest, node.name,
+                         None, None,
+                         None, None,
+                         bls_key=new_blspk,
+                         services=None,
+                         key_proof=None)
+    poolSetExceptOne = list(txnPoolNodeSet)
+    poolSetExceptOne.remove(node)
+    waitNodeDataEquality(looper, node, *poolSetExceptOne)
+    sdk_pool_refresh(looper, sdk_pool_handle)
+    return new_blspk
 
 
 def _create_multi_sig_verifier() -> BlsCryptoVerifier:
